requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Babel==2.0.0
Flask-WTF==1.2.1
python-dotenv==0.19.0
Werkzeug==2.0.1
numpy==1.26.4
matplotlib==3.8.3
pandas==2.2.1
scipy==1.12.0
gunicorn==20.1.0 